refactor: extract TodoList
